move stream related file to script/dom/stream
