+++ components/script/dom/mod.rs
@@ -226,12 +226,8 @@ pub(crate) mod bluetooth;
 #[cfg(feature = "bluetooth")]
 pub(crate) use self::bluetooth::*;
 pub(crate) mod broadcastchannel;
-pub(crate) mod bytelengthqueuingstrategy;
 mod canvas;
 pub(crate) use self::canvas::*;
-pub(crate) mod byteteereadintorequest;
-pub(crate) mod byteteereadrequest;
-pub(crate) mod byteteeunderlyingsource;
 
 pub(crate) mod cdatasection;
 pub(crate) mod characterdata;
@@ -245,7 +241,6 @@ pub(crate) mod comment;
 pub(crate) mod compositionevent;
 pub(crate) mod console;
 pub(crate) mod cookiestore;
-pub(crate) mod countqueuingstrategy;
 mod create;
 pub(crate) mod credentialmanagement;
 pub(crate) use self::credentialmanagement::*;
@@ -264,8 +259,6 @@ pub(crate) mod debuggergetpossiblebreakpointsevent;
 pub(crate) mod debuggerglobalscope;
 pub(crate) mod debuggerpauseevent;
 pub(crate) mod debuggersetbreakpointevent;
-pub(crate) mod defaultteereadrequest;
-pub(crate) mod defaultteeunderlyingsource;
 pub(crate) mod dissimilaroriginlocation;
 pub(crate) mod dissimilaroriginwindow;
 #[expect(dead_code)]
@@ -426,7 +419,6 @@ pub(crate) mod trustedscripturl;
 pub(crate) mod trustedtypepolicy;
 pub(crate) mod trustedtypepolicyfactory;
 pub(crate) mod uievent;
-pub(crate) mod underlyingsourcecontainer;
 pub(crate) mod url;
 pub(crate) mod urlhelper;
 pub(crate) mod urlpattern;
