+++ components/script/dom/stream/defaultteeunderlyingsource.rs
@@ -15,7 +15,7 @@ use super::types::{ReadableStream, ReadableStreamDefaultReader};
 use crate::dom::bindings::error::Error;
 use crate::dom::bindings::reflector::{DomGlobal, Reflector, reflect_dom_object};
 use crate::dom::bindings::root::Dom;
-use crate::dom::defaultteereadrequest::DefaultTeeReadRequest;
+use crate::dom::stream::defaultteereadrequest::DefaultTeeReadRequest;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
 use crate::dom::stream::readablestreamdefaultreader::ReadRequest;
