+++ components/script/dom/stream/mod.rs
@@ -1,5 +1,12 @@
+pub(crate) mod bytelengthqueuingstrategy;
+pub(crate) mod byteteereadintorequest;
+pub(crate) mod byteteereadrequest;
+pub(crate) mod byteteeunderlyingsource;
 pub(crate) mod compressionstream;
+pub(crate) mod countqueuingstrategy;
 pub(crate) mod decompressionstream;
+pub(crate) mod defaultteereadrequest;
+pub(crate) mod defaultteeunderlyingsource;
 pub(crate) mod readablebytestreamcontroller;
 pub(crate) mod readablestream;
 pub(crate) mod readablestreambyobreader;
@@ -9,6 +16,7 @@ pub(crate) mod readablestreamdefaultreader;
 pub(crate) mod readablestreamgenericreader;
 pub(crate) mod transformstream;
 pub(crate) mod transformstreamdefaultcontroller;
+pub(crate) mod underlyingsourcecontainer;
 pub(crate) mod writablestream;
 pub(crate) mod writablestreamdefaultcontroller;
 pub(crate) mod writablestreamdefaultwriter;
