+++ components/script/microtask.rs
@@ -18,9 +18,9 @@ use crate::dom::bindings::cell::DomRefCell;
 use crate::dom::bindings::codegen::Bindings::PromiseBinding::PromiseJobCallback;
 use crate::dom::bindings::codegen::Bindings::VoidFunctionBinding::VoidFunction;
 use crate::dom::bindings::root::DomRoot;
-use crate::dom::byteteereadintorequest::ByteTeeReadIntoRequestMicrotask;
-use crate::dom::byteteereadrequest::ByteTeeReadRequestMicrotask;
-use crate::dom::defaultteereadrequest::DefaultTeeReadRequestMicrotask;
+use crate::dom::stream::byteteereadintorequest::ByteTeeReadIntoRequestMicrotask;
+use crate::dom::stream::byteteereadrequest::ByteTeeReadRequestMicrotask;
+use crate::dom::stream::defaultteereadrequest::DefaultTeeReadRequestMicrotask;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::html::htmlimageelement::ImageElementMicrotask;
 use crate::dom::html::htmlmediaelement::MediaElementMicrotask;
